refactor - get rid of OOP use query
