+++ model/ActivityType.ts
@@ -1,7 +1,4 @@
-import { DB } from "@/utils/db";
+import { DB } from "@/utils/DB";
 
 export class ActivityType {
-    public static async all() {
-        return DB.query('SELECT * FROM activity_types')
-    }
 }
