+++ model/Expense.ts
@@ -1,4 +1,4 @@
-import { DB } from '@/utils/db';
+import { DB } from '@/utils/DB';
 import { Model } from './Core';
 import { Participant } from './Participant';
 
@@ -44,36 +44,5 @@ export class Expense extends Model {
             amount: this.amount,
             date: this.date.toISOString()
         }
-    }
-
-    public async save() {
-        const data = this.toEntity()
-        const columns = Object.keys(data)
-        const values = Object.values(data)
-        const query = `INSERT INTO expenses ( ${columns.join(', ')} ) VALUES ( ${new Array(columns.length).fill("?").join(', ')} )`
-
-        try {
-            return await DB.query(query, values)
-        } catch (e) {
-            if (__DEV__) {
-                console.log(e)
-            }
-            console.error('[INSERT FAIL] - expenses')
-        }
-    }
-
-    public static async get(where?: { [key: string]: [string, string] }): Promise<Expense[]> {
-        let query = "SELECT * FROM expenses";
-        let params: string[] = []
-        if (where) {
-            let temp: string[] = []
-            Object.keys(where).forEach(key => {
-                temp.push(`${key} ${where[key][0]} ?`)
-                params.push(where[key][1])
-            })
-            query = query + " WHERE " + temp.join(' AND ')
-        }
-        const expenses = await DB.query(query, params);
-        return expenses.map((e: any) => new this({ ...e, activityId: e.activity_id, paidBy: e.paid_by }))
     }
 }
